+++ CHANGELOG.md
@@ -44,6 +44,8 @@ When releasing, please add the new-release-boilerplate to docs/pallas/CHANGELOG.
     matrices ({jax-issue}`#10144`).
   * Added {func}`jax.scipy.linalg.helmert` for constructing Helmert matrices
     ({jax-issue}`#10144`).
+  * Added {func}`jax.scipy.special.boxcox` and
+    {func}`jax.scipy.special.boxcox1p` for the Box-Cox power transformation.
   * Moved RNG APIs from "implementations" to dtypes ({jax-issue}`#27854`):
     * Added `jax.random.key_dtype` to get the dtype corresponding to a PRNG
       implementation name.